add routeUsecases files for atomic db operations
