+++ src/usecases/routeUsecases/createRoute.js
@@ -1,0 +1,3 @@
+const createRoute = async (route) => await Route.create(route)
+
+export { createRoute }
