+++ src/usecases/routeUsecases/deleteSingleRoute.js
@@ -1,0 +1,3 @@
+const deleteSingleRoute = async (id) => await Route.findByIdAndDelete(id)
+
+export { deleteSingleRoute }
