+++ src/usecases/routeUsecases/getAllRoutes.js
@@ -1,0 +1,3 @@
+const getAllRoutes = async () => await Route.find()
+
+export { getAllRoutes }
