+++ src/usecases/routeUsecases/getSingleRoute.js
@@ -1,0 +1,3 @@
+const getSingleRoute = async (id) => await Route.findById(id)
+
+export { getSingleRoute }
