+++ src/usecases/routeUsecases/updateSingleRoute.js
@@ -1,0 +1,4 @@
+const updateSingleRoute = async (id, body) =>
+  await Route.findByIdAndUpdate(id, body)
+
+export { updateSingleRoute }
